requirements.txt
aiofiles==24.1.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==41.0.7
fastapi==0.104.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httptools==0.6.4
idna==3.10
packaging==25.0
pefile==2024.8.26
pycparser==2.23
pydantic==2.5.0
pydantic_core==2.14.1
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2025.8
python-dotenv==1.1.1
python-multipart==0.0.6
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.31.0
setuptools==80.9.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
websockets==12.0
